Separate parsing logic
